// src/modules/excel-reports/domain/enums/facturacion-report/facturacion-reporte-estado-factura.enum.ts
export enum FacturacionReporteEstadoFactura {
  PENDIENTE = 'PENDIENTE',
  PAGADA = 'PAGADA',
  VENCIDA = 'VENCIDA',
  ANULADA = 'ANULADA',
  PARCIAL = 'PARCIAL',
}

// export type FacturacionReporteEstadoFactura =
//   (typeof FacturacionReporteEstadoFactura)[keyof typeof FacturacionReporteEstadoFactura];