// src/modules/excel-reports/domain/enums/facturacion-report/facturacion-reporte-metodo-pago.enum.ts
export enum FacturacionReporteMetodoPago {
  EFECTIVO = 'EFECTIVO',
  TARJETA = 'TARJETA',
  DEPOSITO = 'DEPOSITO',
  PAYPAL = 'PAYPAL',
  PENDIENTE = 'PENDIENTE',
  OTRO = 'OTRO',
}

// export type FacturacionReporteMetodoPago =
//   (typeof FacturacionReporteMetodoPago)[keyof typeof FacturacionReporteMetodoPago];